requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openlibx402-core==0.1.0
openlibx402-fastapi==0.1.0
python-dotenv==1.0.0
solders==0.21.0
solana==0.34.3
